Add files via upload
The SQL file analyzes COVID-19 data to explore infection and mortality rates. It includes queries to calculate death percentages per case, the proportion of the population infected, and identify the most affected countries and continents. The goal is to gain insights into the virus's spread and impact by population and region.